requirements.txt
beautifulsoup4>=4.9.3
requests>=2.25.1
streamlit>=1.10.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
python-dotenv>=0.19.0
retry>=0.9.2